requirements.txt
antiorm==1.2.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
db==0.1.1
db-sqlite3==0.0.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-guardian==2.3.0
django-import-export==2.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
twilio==6.50.1
urllib3==1.26.2
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
